optimized top-level python layout
